requirements.txt
streamlit
opencv-python
scikit-learn
torch
torchvision
numpy
Pillow
matplotlib
pandas